refactor(client): move wrapper into index and rename dumb component into view to follow naming conventions
