--- client/src/Main/Monitor/ResultPanel/ResultPanel.wrap.js
@@ -1,13 +1,0 @@
-import { connect } from 'react-redux';
-import ResultPanel from './ResultPanel';
-import { setTaskFields } from '../MonitorActions';
-
-const mapStateToProps = state => ({});
-
-const mapDispatchToProps = dispatch => ({
-  handleTaskChange: (taskIndex, fields) => {
-    dispatch(setTaskFields(taskIndex, fields));
-  },
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(ResultPanel);
