+++ client/src/Main/Monitor/ResultPanel/index.js
@@ -1,1 +1,13 @@
-export { default } from './ResultPanel.wrap';
+import { connect } from 'react-redux';
+import ResultPanel from './view';
+import { setTaskFields } from '../MonitorActions';
+
+const mapStateToProps = state => ({});
+
+const mapDispatchToProps = dispatch => ({
+  handleTaskChange: (taskIndex, fields) => {
+    dispatch(setTaskFields(taskIndex, fields));
+  },
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(ResultPanel);
