+++ client/src/Main/Monitor/ResultPanel/view.js
